chore: fix cached fetch
