+++ engines/query-sparql/test/QuerySparql-test.ts
@@ -1,6 +1,5 @@
 /** @jest-environment setup-polly-jest/jest-environment-node */
-import * as fs from 'node:fs';
-import * as path from 'node:path';
+
 import { QuerySourceSkolemized } from '@comunica/actor-context-preprocess-query-source-skolemize';
 import { KeysHttpWayback, KeysQuerySourceIdentify } from '@comunica/context-entries';
 import { BlankNodeScoped } from '@comunica/data-factory';
@@ -12,8 +11,7 @@ import { Store } from 'n3';
 import { DataFactory } from 'rdf-data-factory';
 import { Factory } from 'sparqlalgebrajs';
 import { QueryEngine } from '../lib/QueryEngine';
-
-const md5 = require('md5');
+import { fetch as cachedFetch } from './util';
 
 // Use require instead of import for default exports, to be compatible with variants of esModuleInterop in tsconfig.
 const stringifyStream = require('stream-to-string');
@@ -21,30 +19,11 @@ const stringifyStream = require('stream-to-string');
 const DF = new DataFactory();
 const factory = new Factory();
 
-const fetchFn = globalThis.fetch;
-globalThis.fetch = async(...args: Parameters<typeof fetch>): ReturnType<typeof fetch> => {
-  // eslint-disable-next-line ts/no-base-to-string
-  const pth = path.join(__dirname, 'networkCache', md5(args[0].toString()));
-  if (!fs.existsSync(pth)) {
-    const res = await fetchFn(...args);
-    const headersObject: Record<string, string> = {};
-    for (const [ key, value ] of res.headers.entries()) {
-      headersObject[key] = value;
-    }
-    fs.writeFileSync(pth, JSON.stringify({
-      content: await res.text().catch(() => ''),
-      headers: headersObject,
-      status: res.status,
-      statusText: res.statusText,
-    }));
-  }
-  const { content, headers, status, statusText } = JSON.parse(fs.readFileSync(pth).toString());
-  return new Response(content, { headers, status, statusText });
-};
+globalThis.fetch = cachedFetch;
 
 describe('System test: QuerySparql', () => {
   let engine: QueryEngine;
-  beforeAll(() => {
+  beforeEach(() => {
     engine = new QueryEngine();
   });
 
@@ -117,23 +96,21 @@ describe('System test: QuerySparql', () => {
 
       describe('string source query', () => {
         const query = 'CONSTRUCT WHERE { ?s ?p ?o }';
-        const value = `{
-          "@id":"ex:s",
-          "ex:p":{"@id":"ex:o"},
-          "ex:p2":{"@id":"ex:o2"}
-        }
-        `;
-        const value2 = '<ex:s> <ex:p3> <ex:o3>. <ex:s> <ex:p4> <ex:o4>.';
-        const expectedResult: RDF.Quad[] = [
-          DF.quad(DF.namedNode('ex:s'), DF.namedNode('ex:p'), DF.namedNode('ex:o')),
-          DF.quad(DF.namedNode('ex:s'), DF.namedNode('ex:p2'), DF.namedNode('ex:o2')),
-        ];
+
+        beforeEach(() => {
+          engine = new QueryEngine();
+        });
 
         it('should return the valid result with a turtle data source', async() => {
-          const turtleValue = '<ex:s> <ex:p> <ex:o>. <ex:s> <ex:p2> <ex:o2>.';
+          const value = '<ex:s> <ex:p> <ex:o>. <ex:s> <ex:p2> <ex:o2>.';
           const context: QueryStringContext = { sources: [
-            { type: 'serialized', value: turtleValue, mediaType: 'text/turtle', baseIRI: 'http://example.org/' },
+            { type: 'serialized', value, mediaType: 'text/turtle', baseIRI: 'http://example.org/' },
           ]};
+
+          const expectedResult: RDF.Quad[] = [
+            DF.quad(DF.namedNode('ex:s'), DF.namedNode('ex:p'), DF.namedNode('ex:o')),
+            DF.quad(DF.namedNode('ex:s'), DF.namedNode('ex:p2'), DF.namedNode('ex:o2')),
+          ];
 
           const result = await arrayifyStream(await engine.queryQuads(query, context));
           expect(result).toHaveLength(expectedResult.length);
@@ -141,9 +118,20 @@ describe('System test: QuerySparql', () => {
         });
 
         it('should return the valid result with a json-ld data source', async() => {
+          const value = `{
+            "@id":"ex:s",
+            "ex:p":{"@id":"ex:o"},
+            "ex:p2":{"@id":"ex:o2"}
+          }
+          `;
           const context: QueryStringContext = { sources: [
             { type: 'serialized', value, mediaType: 'application/ld+json', baseIRI: 'http://example.org/' },
           ]};
+
+          const expectedResult: RDF.Quad[] = [
+            DF.quad(DF.namedNode('ex:s'), DF.namedNode('ex:p'), DF.namedNode('ex:o')),
+            DF.quad(DF.namedNode('ex:s'), DF.namedNode('ex:p2'), DF.namedNode('ex:o2')),
+          ];
 
           const result = await arrayifyStream(await engine.queryQuads(query, context));
           expect(result).toHaveLength(expectedResult.length);
@@ -151,9 +139,19 @@ describe('System test: QuerySparql', () => {
         });
 
         it('should return the valid result with no base IRI', async() => {
+          const value = `{
+            "@id":"ex:s",
+            "ex:p":{"@id":"ex:o"},
+            "ex:p2":{"@id":"ex:o2"}
+          }`;
           const context: QueryStringContext = { sources: [
             { type: 'serialized', value, mediaType: 'application/ld+json' },
           ]};
+
+          const expectedResult: RDF.Quad[] = [
+            DF.quad(DF.namedNode('ex:s'), DF.namedNode('ex:p'), DF.namedNode('ex:o')),
+            DF.quad(DF.namedNode('ex:s'), DF.namedNode('ex:p2'), DF.namedNode('ex:o2')),
+          ];
 
           const result = await arrayifyStream(await engine.queryQuads(query, context));
           expect(result).toHaveLength(expectedResult.length);
@@ -161,8 +159,14 @@ describe('System test: QuerySparql', () => {
         });
 
         it('should return the valid result with multiple serialized', async() => {
+          const value1 = `{
+            "@id":"ex:s",
+            "ex:p":{"@id":"ex:o"},
+            "ex:p2":{"@id":"ex:o2"}
+          }`;
+          const value2 = '<ex:s> <ex:p3> <ex:o3>. <ex:s> <ex:p4> <ex:o4>.';
           const context: QueryStringContext = { sources: [
-            { type: 'serialized', value, mediaType: 'application/ld+json' },
+            { type: 'serialized', value: value1, mediaType: 'application/ld+json' },
             { type: 'serialized', value: value2, mediaType: 'text/turtle' },
           ]};
 
@@ -185,8 +189,14 @@ describe('System test: QuerySparql', () => {
           const store = new Store();
           store.addQuads(quads);
 
+          const value1 = `{
+            "@id":"ex:s",
+            "ex:p":{"@id":"ex:o"},
+            "ex:p2":{"@id":"ex:o2"}
+          }`;
+          const value2 = '<ex:s> <ex:p3> <ex:o3>. <ex:s> <ex:p4> <ex:o4>.';
           const context: QueryStringContext = { sources: [
-            { type: 'serialized', value, mediaType: 'application/ld+json' },
+            { type: 'serialized', value: value1, mediaType: 'application/ld+json' },
             { type: 'serialized', value: value2, mediaType: 'text/turtle' },
             store,
           ]};
@@ -198,32 +208,6 @@ describe('System test: QuerySparql', () => {
             DF.quad(DF.namedNode('ex:s'), DF.namedNode('ex:p4'), DF.namedNode('ex:o4')),
           ]);
         });
-
-        it('should serialise results correctly', async() => {
-          const context: QueryStringContext = { sources: [
-            { type: 'serialized', value, mediaType: 'application/ld+json', baseIRI: 'http://example.org/' },
-          ]};
-
-          const resultString = await engine.resultToString(await engine.query(query, context), 'application/n-quads');
-          await expect(stringifyStream(resultString.data)).resolves.toBe(
-            '<ex:s> <ex:p> <ex:o> .\n<ex:s> <ex:p2> <ex:o2> .\n',
-          );
-        });
-
-        it('should have stats that are strictly increasing', async() => {
-          const context: QueryStringContext = { sources: [
-            { type: 'serialized', value, mediaType: 'application/ld+json', baseIRI: 'http://example.org/' },
-          ]};
-
-          const resultString: string = await stringifyStream(
-            (await engine.resultToString(await engine.query(query, context), 'stats')).data,
-          );
-          const times = resultString.split('\n').slice(1, -1).map(line => Number.parseFloat(line.split(',')[1]));
-          expect(times).toHaveLength(3);
-          for (let i = 0; i < 2; i++) {
-            expect(times[i]).toBeLessThanOrEqual(times[i + 1]);
-          }
-        });
       });
 
       describe('handle blank nodes with DESCRIBE queries', () => {
@@ -234,6 +218,7 @@ describe('System test: QuerySparql', () => {
       }`;
 
         beforeEach(() => {
+          engine = new QueryEngine();
           store = new Store();
         });
