+++ engines/query-sparql/test/util-browser.js
@@ -1,3 +1,5 @@
-export function usePolly() {
-    // No-op
+const fetchFn = globalThis.fetch;
+
+export function fetch(...args) {
+    return fetchFn(...args);
 }
